requirements.txt
Django==2.2.12
django-appconf==1.0.4
django-ckeditor==5.9.0
django-geojson==3.0.0
django-imagekit==4.0.2
django-js-asset==1.2.2
django-leaflet==0.26.0
django-recaptcha==2.0.6
jsonfield==3.0.0
pilkit==2.0
Pillow==7.1.1
python-magic==0.4.15
pytz==2019.3
six==1.14.0
sqlparse==0.3.1
